fix: 테스트 메서드 위치 변경
